requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
environs==10.3.0
marshmallow==3.20.2
packaging==23.2
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
